move common methods from CategoryStratum to Stratum
