--- client/lib/CategoryStratum/getNodes.js
@@ -1,8 +1,0 @@
-module.exports = function (nodeKey) {
-  // Get nodes in an array.
-  //
-  // Return
-  //   array of CategoryNode
-  //
-  return Object.values(this.renderedNodes)
-}
