+++ client/lib/Stratum/getBoundingCircle.js
@@ -1,4 +1,6 @@
 module.exports = function () {
+  // @Stratum:getBoundingCircle()
+  //
   // Get the cached stratum boundary.
   //
   // Return
