+++ client/lib/Stratum/getNodes.js
@@ -1,0 +1,8 @@
+module.exports = function () {
+  // Get stratum nodes in an array.
+  //
+  // Return
+  //   array of StratumNode
+  //
+  return Object.values(this.renderedNodes)
+}
