+++ client/lib/Stratum/recomputeBoundingCircle.js
